Merge ABI break into master

+++ options/posix/include/sys/resource.h
@@ -25,7 +25,7 @@
 extern "C" {
 #endif
 
-typedef unsigned long rlim_t;
+typedef unsigned long long rlim_t;
 
 struct rlimit {
 	rlim_t rlim_cur;
